NICK CHECK

+++ Database_channels.cpp
@@ -53,7 +53,7 @@ void	Database::delete_channel(std::string& chan_name)
 }
 
 /// @brief sender, command -> error_message에 보낼 정보
-Channel&	Database::select_channel(std::string& chan_name, int error_code, User& sender, std::string& command) // 403 ERROR Sender::no_channel_message
+Channel&	Database::select_channel(std::string& chan_name) // 403 ERROR Sender::no_channel_message
 {
 	std::vector<Channel>::iterator it = Channels_.begin();
 	for (; it != Channels_.end(); ++it)
@@ -63,11 +63,10 @@ Channel&	Database::select_channel(std::string& chan_name, int error_code, User& 
 			return *it;
 		}
 	}
-	//Sender::error_message(sender.client_sock_, command, error_code);
 	return *it;
 }
 
-Channel&	Database::select_channel(User& connector, int error_code, std::string& command) // 476 ERROR Sender::join_invaild_channel_name_message
+Channel&	Database::select_channel(User& connector) // 476 ERROR Sender::join_invaild_channel_name_message
 {
 	std::vector<Channel>::iterator it = Channels_.begin();
 	for (; it != Channels_.end(); ++it)
@@ -77,7 +76,6 @@ Channel&	Database::select_channel(User& connector, int error_code, std::string& 
 			return *it;
 		}
 	}
-	//Sender::error_message(connector.client_sock_, command, error_code);
 	return *it;
 }
 
@@ -86,9 +84,14 @@ Udata	Database::join_channel(User& joiner, std::string& chan_name)
 {
 	Udata				ret;
 	Event				tmp;
-	std::string			command("JOIN");
 
-	Channel& chan = select_channel(chan_name, 476, joiner, command);
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(joiner, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel& chan = select_channel(chan_name);
 	if (chan.is_user(joiner) == false)
 	{
 		tmp = Sender::no_user_message(joiner, chan_name);
@@ -104,27 +107,32 @@ Udata	Database::join_channel(User& joiner, std::string& chan_name)
 		it->second += Sender::join_353_message(joiner, chan.get_name(), chan.get_access(), chan_user_list);
 		it->second += Sender::join_366_message(joiner, chan.get_name());
 	}
-
 	return ret;
 }
 
 Udata	Database::leave_channel(User&leaver, std::string& chan_name, std::string& msg)
 {
 	Event				tmp;
-	Udata				res;
-	std::string			command("PART");
+	Udata				ret;
 
-	Channel& chan = select_channel(chan_name, 403, leaver, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(leaver, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel& chan = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	//유저가 존재하지 않을 경우(로비에서 part하면 예외처리)
 	if (chan.is_user(leaver) == 0)
 	{
-		return res;
+		tmp = Sender::no_user_message(leaver, leaver.nickname_);
+		return ret;
 	}
 	//Msg전송 : PART 내용에 따라 전송 -> 아마 채널의 다른 유저들에게 떠났다고 알려줘야
 	std::vector<User> users = chan.get_users();
 	int user_size = users.size();
 	//PART하면, 그 내역은 모두에게 보내진다. 나간 사람 포함한다.
-	res = chan.send_all(leaver, leaver, msg, PART);
+	ret = chan.send_all(leaver, leaver, msg, PART);
 	chan.delete_user(leaver);
 
 	if (user_size == 1) //채널 잘 삭제되는 것 확인
@@ -138,7 +146,7 @@ Udata	Database::leave_channel(User&leaver, std::string& chan_name, std::string& 
 			chan.set_host();
 		}
 	}
-	return res;
+	return ret;
 }
 
 /// @brief 채널 전체 유저에게 메시지 전달. 내외부 모두 사용됨
@@ -146,9 +154,14 @@ Udata	Database::channel_msg(User& sender, std::string chan_name, std::string& ms
 {
 	Udata		ret;
 	Event		tmp;
-	std::string	command("PRIVMSG");
 
-	Channel&	channel = select_channel(chan_name, 403, sender, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(sender, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel&	channel = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	ret = channel.send_all(sender, sender, msg, PRIV);
 	return ret;
 }
@@ -157,9 +170,14 @@ Udata	Database::channel_notice(User& sender, std::string chan_name, std::string&
 {
 	Udata			ret;
 	Event			tmp;
-	std::string		command("NOTICE");
 
-	Channel&	channel = select_channel(chan_name, 403, sender, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(sender, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel&	channel = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	ret = channel.send_all(sender, sender, msg, NOTICE);
 	return ret;
 }
@@ -168,11 +186,15 @@ Udata	Database::channel_wall(User& sender, std::string chan_name, std::string& m
 {
 	Udata			ret;
 	Event			tmp;
-	std::string		command("WALL");
 
-	Channel&	channel = select_channel(chan_name, 403, sender, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(sender, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel&	channel = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	User	host = channel.get_host();
-
 	if (host == sender)
 	{
 		std::cout << "[Debug] " << "host == sender" << std::endl;
@@ -187,10 +209,14 @@ Udata	Database::kick_channel(User& host, User& target, std::string& chan_name, s
 {
 	Udata			ret;
 	Event			tmp;
-	std::string		command("PRIVMSG");
 
-
-	Channel&	channel = select_channel(chan_name, 403, host, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(host, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel&	channel = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	if (channel.get_host() == host)
 	{
 		if (channel.is_user(target) == true)
@@ -217,10 +243,14 @@ Udata	Database::who_channel(User& asker, std::string& chan_name)
 {
 	Udata	ret;
 	Event	tmp;
-	std::string command("WHO");
 
-	Channel& chan = select_channel(chan_name, 315, asker, command); // 315 -> end of list만 보낸다.
-
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(asker, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel& chan = select_channel(chan_name); // 315 -> end of list만 보낸다.
 	tmp = Sender::who_joiner_352_message(asker, chan_name);
 	tmp.second += Sender::who_315_message(asker, chan_name);
 
@@ -232,16 +262,21 @@ Udata	Database::quit_channel(User& leaver, std::string msg)
 {
 	Udata			ret;
 	Event			tmp;
-	std::string		command("QUIT");
 
-	Channel& channel = select_channel(leaver, 401, command); // 401 no_such_nick
+	if (is_user_in_channel(leaver) == false)
+	{
+		tmp = Sender::no_user_message(leaver, leaver.nickname_);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel& channel = select_channel(leaver); // 401 no_such_nick
 	ret = channel.send_all(leaver, leaver, msg, QUIT);
 	channel.delete_user(leaver);
 	
 	return ret;
 }
 
-Udata	Database::mode_channel(User& moder, const std::string& chan_name, const bool vaild)
+Udata	Database::mode_channel(User& moder, std::string& chan_name, bool vaild)
 {
 	Udata	ret;
 	Event	tmp;
@@ -250,6 +285,13 @@ Udata	Database::mode_channel(User& moder, const std::string& chan_name, const bo
 	std::stringstream	ss;
 	ss << result;
 	const std::string& time_stamp(ss.str());
+
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(moder, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
 
 	if (vaild)	// true
 	{
@@ -261,16 +303,15 @@ Udata	Database::mode_channel(User& moder, const std::string& chan_name, const bo
 	return ret;
 }
 
-Udata	Database::nick_channel(User& nicker, std::string& send_msg)
+Udata	Database::nick_channel(User& nicker, std::string& new_nick)
 {
 	Udata		ret;
-	User	trash;
-	std::string command("NICK");
-	
-	Channel& channel = select_channel(nicker, 401, command); // 401 no such nick
+	Event		tmp;
+	User		trash;
 
-	channel.change_nick(nicker, send_msg);
-	ret = channel.send_all(nicker, trash, send_msg, NICK);
+	Channel& channel = select_channel(nicker); // 401 no such nick
+	channel.change_nick(nicker, new_nick);
+	ret = channel.send_all(nicker, trash, new_nick, NICK);
 	return ret;
 }
 
@@ -278,10 +319,14 @@ Udata	Database::set_topic(User& sender, std::string& chan_name, std::string& top
 {
 	Udata			ret;
 	Event			tmp;
-	std::string		command("TOPIC");
 
-
-	Channel& channel = select_channel(chan_name, 403, sender, command); // 403 ERROR Sender::no_channel_message
+	if (is_channel(chan_name) == false)
+	{
+		tmp = Sender::no_channel_message(sender, chan_name);
+		ret.insert(tmp);
+		return ret;
+	}
+	Channel& channel = select_channel(chan_name); // 403 ERROR Sender::no_channel_message
 	if (channel.get_host() == sender)
 	{
 		std::string topic_msg = "Topic was changed to " + topic;
